requirements.txt
asgiref==3.2.10
azure-core==1.13.0
azure-storage-blob==12.8.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
crypto==1.4.1
cryptography==3.4.7
Django==3.1
django-crispy-forms==1.11.2
django-filter==2.4.0
djangorestframework==3.12.4
drf-yasg==1.20.0
Faker==8.8.2
# h11 @ file:///tmp/build/80754af9/h11_1620423447028/work
h2==3.2.0
hpack==3.0.0
# httpcore @ file:///tmp/build/80754af9/httpcore_1620428819128/work
# httpx @ file:///tmp/build/80754af9/httpx_1624368550453/work
hyperframe==5.2.0
# idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
inflection==0.5.1
isodate==0.6.0
itypes==1.2.0
Jinja2==3.0.1
mailer==0.8.1
mailjet-rest==1.3.4
Markdown==3.3.4
MarkupSafe==2.0.1
msrest==0.6.21
Naked==0.1.31
numpy==1.21.0
oauthlib==3.1.0
packaging==20.9
pandas==1.2.5
Pillow==8.0.0
psycopg2-binary==2.9.1
pyasn1==0.4.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
# pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
# rfc3986 @ file:///tmp/build/80754af9/rfc3986_1594058972433/work
rsa==4.7.2
ruamel.yaml==0.17.9
ruamel.yaml.clib==0.2.2
shellescape==3.8.1
six==1.16.0
# sniffio @ file:///C:/ci/sniffio_1614030707456/work
# sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.4
wincertstore==0.2